[Cleanup] Move structs to separate .h/.cpp files
