+++ src/DataStructs/SecurityStruct.cpp
@@ -1,0 +1,32 @@
+#include "DataStructs/SecurityStruct.h"
+
+#include "ESPEasy_common.h"
+#include "DataStructs/ESPEasyLimits.h"
+
+SecurityStruct::SecurityStruct() {
+  ZERO_FILL(WifiSSID);
+  ZERO_FILL(WifiKey);
+  ZERO_FILL(WifiSSID2);
+  ZERO_FILL(WifiKey2);
+  ZERO_FILL(WifiAPKey);
+
+  for (byte i = 0; i < CONTROLLER_MAX; ++i) {
+    ZERO_FILL(ControllerUser[i]);
+    ZERO_FILL(ControllerPassword[i]);
+  }
+  ZERO_FILL(Password);
+}
+
+void SecurityStruct::validate() {
+  ZERO_TERMINATE(WifiSSID);
+  ZERO_TERMINATE(WifiKey);
+  ZERO_TERMINATE(WifiSSID2);
+  ZERO_TERMINATE(WifiKey2);
+  ZERO_TERMINATE(WifiAPKey);
+
+  for (byte i = 0; i < CONTROLLER_MAX; ++i) {
+    ZERO_TERMINATE(ControllerUser[i]);
+    ZERO_TERMINATE(ControllerPassword[i]);
+  }
+  ZERO_TERMINATE(Password);
+}
