+++ src/ESPEasy_common.cpp
@@ -1,5 +1,8 @@
 #include "ESPEasy_common.h"
 
+
+String getUnknownString() { return F("Unknown"); }
+
 /*********************************************************************************************\
    Bitwise operators
   \*********************************************************************************************/
