+++ src/ESPEasy_common.h
@@ -15,6 +15,8 @@ namespace std
 #define ZERO_TERMINATE(S)  S[sizeof(S) - 1] = 0
 
 
+String getUnknownString();
+
 /*********************************************************************************************\
    Bitwise operators
   \*********************************************************************************************/
