+++ src/_CPlugin_Helper.cpp
@@ -1,0 +1,422 @@
+#include "_CPlugin_Helper.h"
+
+#include "ESPEasy_fdwdecl.h"
+#include "ESPEasy_Log.h"
+#include "ESPEasy_buildinfo.h"
+
+#include "DataStructs/SecurityStruct.h"
+#include "DataStructs/CRCStruct.h"
+#include "DataStructs/SettingsStruct.h"
+
+#include "DataStructs/ControllerSettingsStruct.h"
+#include "DataStructs/ESPEasyLimits.h"
+#include "DataStructs/TimingStats.h"
+
+#include <WiFiClient.h>
+#include <WiFiUdp.h>
+#include <base64.h>
+
+bool safeReadStringUntil(Stream     & input,
+                         String     & str,
+                         char         terminator,
+                         unsigned int maxSize,
+                         unsigned int timeout)
+{
+  int c;
+  const unsigned long start           = millis();
+  const unsigned long timer           = start + timeout;
+  unsigned long backgroundtasks_timer = start + 10;
+
+  str = "";
+
+  do {
+    // read character
+    if (input.available()) {
+      c = input.read();
+
+      if (c >= 0) {
+        // found terminator, we're ok
+        if (c == terminator) {
+          return true;
+        }
+
+        // found character, add to string
+        else {
+          str += char(c);
+
+          // string at max size?
+          if (str.length() >= maxSize) {
+            addLog(LOG_LEVEL_ERROR, F("Not enough bufferspace to read all input data!"));
+            return false;
+          }
+        }
+      }
+
+      // We must run the backgroundtasks every now and then.
+      if (timeOutReached(backgroundtasks_timer)) {
+        backgroundtasks_timer += 10;
+        backgroundtasks();
+      } else {
+        delay(0);
+      }
+    } else {
+      delay(0);
+    }
+  } while (!timeOutReached(timer));
+
+  addLog(LOG_LEVEL_ERROR, F("Timeout while reading input data!"));
+  return false;
+}
+
+bool valid_controller_number(int controller_number) {
+  if (controller_number < 0) { return false; }
+  return true;
+
+  //  return getProtocolIndex(controller_number) <= protocolCount;
+}
+
+String get_formatted_Controller_number(int controller_number) {
+  if (!valid_controller_number(controller_number)) {
+    return F("C---");
+  }
+  String result = F("C");
+
+  if (controller_number < 100) { result += '0'; }
+
+  if (controller_number < 10) { result += '0'; }
+  result += controller_number;
+  return result;
+}
+
+
+String get_auth_header(const String& user, const String& pass) {
+  String authHeader = "";
+
+  if (user.length() != 0 && pass.length() != 0) {
+      base64 encoder;
+      String auth = user;
+      auth       += ":";
+      auth       += pass;
+      authHeader  = F("Authorization: Basic ");
+      authHeader += encoder.encode(auth);
+      authHeader += F(" \r\n");
+  }
+  return authHeader;
+}
+
+String get_auth_header(int controller_index) {
+  String authHeader = "";
+
+  if (controller_index < CONTROLLER_MAX) {
+    if ((getSecuritySettings().ControllerUser[controller_index][0] != 0) &&
+        (getSecuritySettings().ControllerPassword[controller_index][0] != 0))
+    {
+      authHeader = get_auth_header(
+        String(getSecuritySettings().ControllerUser[controller_index]), 
+        String(getSecuritySettings().ControllerPassword[controller_index]));
+    }
+  } else {
+    addLog(LOG_LEVEL_ERROR, F("Invalid controller index"));
+  }
+  return authHeader;
+}
+
+String get_user_agent_request_header_field() {
+  static unsigned int agent_size = 20;
+  String request;
+
+  request.reserve(agent_size);
+  request    = F("User-Agent: ");
+  request   += F("ESP Easy/");
+  request   += BUILD;
+  request   += '/';
+  request   += String(getCRCValues().compileDate);
+  request   += ' ';
+  request   += String(getCRCValues().compileTime);
+  request   += "\r\n";
+  agent_size = request.length();
+  return request;
+}
+
+String do_create_http_request(
+  const String& hostportString,
+  const String& method, const String& uri,
+  const String& auth_header, const String& additional_options,
+  int content_length) {
+  int estimated_size = hostportString.length() + method.length()
+                       + uri.length() + auth_header.length()
+                       + additional_options.length()
+                       + 42;
+
+  if (content_length >= 0) { estimated_size += 25; }
+  String request;
+  request.reserve(estimated_size);
+  request += method;
+  request += ' ';
+
+  if (!uri.startsWith("/")) { request += '/'; }
+  request += uri;
+  request += F(" HTTP/1.1");
+  request += "\r\n";
+
+  if (content_length >= 0) {
+    request += F("Content-Length: ");
+    request += content_length;
+    request += "\r\n";
+  }
+  request += F("Host: ");
+  request += hostportString;
+  request += "\r\n";
+  request += auth_header;
+  request += additional_options;
+  request += get_user_agent_request_header_field();
+  request += F("Connection: close\r\n");
+  request += "\r\n";
+#ifndef BUILD_NO_DEBUG
+  addLog(LOG_LEVEL_DEBUG, request);
+#endif // ifndef BUILD_NO_DEBUG
+  return request;
+}
+
+String do_create_http_request(
+  const String& hostportString,
+  const String& method, const String& uri) {
+  return do_create_http_request(hostportString, method, uri,
+                                "", // auth_header
+                                "", // additional_options
+                                -1  // content_length
+                                );
+}
+
+String do_create_http_request(
+  int controller_number, ControllerSettingsStruct& ControllerSettings,
+  const String& method, const String& uri,
+  int content_length) {
+  const bool defaultport = ControllerSettings.Port == 0 || ControllerSettings.Port == 80;
+
+  return do_create_http_request(
+    defaultport ? ControllerSettings.getHost() : ControllerSettings.getHostPortString(),
+    method,
+    uri,
+    "", // auth_header
+    "", // additional_options
+    content_length);
+}
+
+String create_http_request_auth(
+  int controller_number, int controller_index, ControllerSettingsStruct& ControllerSettings,
+  const String& method, const String& uri,
+  int content_length) {
+  const bool defaultport = ControllerSettings.Port == 0 || ControllerSettings.Port == 80;
+
+  return do_create_http_request(
+    defaultport ? ControllerSettings.getHost() : ControllerSettings.getHostPortString(),
+    method,
+    uri,
+    get_auth_header(controller_index),
+    "", // additional_options
+    content_length);
+}
+
+String create_http_get_request(int controller_number, ControllerSettingsStruct& ControllerSettings,
+                               const String& uri) {
+  return do_create_http_request(controller_number, ControllerSettings, F("GET"), uri, -1);
+}
+
+String create_http_request_auth(int controller_number, int controller_index, ControllerSettingsStruct& ControllerSettings,
+                                const String& method, const String& uri) {
+  return create_http_request_auth(controller_number, controller_index, ControllerSettings, method, uri, -1);
+}
+
+#ifndef BUILD_NO_DEBUG
+void log_connecting_to(const String& prefix, int controller_number, ControllerSettingsStruct& ControllerSettings) {
+  if (loglevelActiveFor(LOG_LEVEL_DEBUG)) {
+    String log = prefix;
+    log += get_formatted_Controller_number(controller_number);
+    log += F(" connecting to ");
+    log += ControllerSettings.getHostPortString();
+    addLog(LOG_LEVEL_DEBUG, log);
+  }
+}
+
+#endif // ifndef BUILD_NO_DEBUG
+
+void log_connecting_fail(const String& prefix, int controller_number, ControllerSettingsStruct& ControllerSettings) {
+  if (loglevelActiveFor(LOG_LEVEL_ERROR)) {
+    String log = prefix;
+    log += get_formatted_Controller_number(controller_number);
+    log += F(" connection failed (");
+    log += getConnectionFailures();
+    log += F("/");
+    log += getSettings().ConnectionFailuresThreshold;
+    log += F(")");
+    addLog(LOG_LEVEL_ERROR, log);
+  }
+}
+
+bool count_connection_results(bool success, const String& prefix, int controller_number, ControllerSettingsStruct& ControllerSettings) {
+  if (!success)
+  {
+    getConnectionFailures()++;
+    log_connecting_fail(prefix, controller_number, ControllerSettings);
+    return false;
+  }
+  statusLED(true);
+
+  if (getConnectionFailures()) {
+    getConnectionFailures()--;
+  }
+  return true;
+}
+
+bool try_connect_host(int controller_number, WiFiUDP& client, ControllerSettingsStruct& ControllerSettings) {
+  //START_TIMER; // FIXME TD-er Timingstats macros currently not callable from .cpp file
+  if (!WiFiConnected()) return false;
+  client.setTimeout(ControllerSettings.ClientTimeout);
+#ifndef BUILD_NO_DEBUG
+  log_connecting_to(F("UDP  : "), controller_number, ControllerSettings);
+#endif // ifndef BUILD_NO_DEBUG
+  bool success      = ControllerSettings.beginPacket(client) != 0;
+  const bool result = count_connection_results(
+    success,
+    F("UDP  : "), controller_number, ControllerSettings);
+  //STOP_TIMER(TRY_CONNECT_HOST_UDP);
+  return result;
+}
+
+bool try_connect_host(int controller_number, WiFiClient& client, ControllerSettingsStruct& ControllerSettings) {
+  //START_TIMER; // FIXME TD-er Timingstats macros currently not callable from .cpp file
+  if (!WiFiConnected()) return false;
+
+  // Use WiFiClient class to create TCP connections
+  client.setTimeout(ControllerSettings.ClientTimeout);
+#ifndef BUILD_NO_DEBUG
+  log_connecting_to(F("HTTP : "), controller_number, ControllerSettings);
+#endif // ifndef BUILD_NO_DEBUG
+  bool success      = ControllerSettings.connectToHost(client);
+  const bool result = count_connection_results(
+    success,
+    F("HTTP : "), controller_number, ControllerSettings);
+  //STOP_TIMER(TRY_CONNECT_HOST_TCP);
+  return result;
+}
+
+// Use "client.available() || client.connected()" to read all lines from slow servers.
+// See: https://github.com/esp8266/Arduino/pull/5113
+//      https://github.com/esp8266/Arduino/pull/1829
+bool client_available(WiFiClient& client) {
+  delay(0);
+  return client.available() || client.connected();
+}
+
+bool send_via_http(const String& logIdentifier, WiFiClient& client, const String& postStr, bool must_check_reply) {
+  bool success = !must_check_reply;
+
+  // This will send the request to the server
+  byte written = client.print(postStr);
+
+  // as of 2018/11/01 the print function only returns one byte (upd to 256 chars sent). However if the string sent can be longer than this
+  // therefore we calculate modulo 256.
+  // see discussion here https://github.com/letscontrolit/ESPEasy/pull/1979
+  // and implementation here
+  // https://github.com/esp8266/Arduino/blob/561426c0c77e9d05708f2c4bf2a956d3552a3706/libraries/ESP8266WiFi/src/include/ClientContext.h#L437-L467
+  // this needs to be adjusted if the WiFiClient.print method changes.
+  if (written != (postStr.length() % 256)) {
+    if (loglevelActiveFor(LOG_LEVEL_ERROR)) {
+      String log = F("HTTP : ");
+      log += logIdentifier;
+      log += F(" Error: could not write to client (");
+      log += written;
+      log += "/";
+      log += postStr.length();
+      log += ")";
+      addLog(LOG_LEVEL_ERROR, log);
+    }
+    success = false;
+  }
+#ifndef BUILD_NO_DEBUG
+  else {
+    if (loglevelActiveFor(LOG_LEVEL_DEBUG)) {
+      String log = F("HTTP : ");
+      log += logIdentifier;
+      log += F(" written to client (");
+      log += written;
+      log += "/";
+      log += postStr.length();
+      log += ")";
+      addLog(LOG_LEVEL_DEBUG, log);
+    }
+  }
+#endif // ifndef BUILD_NO_DEBUG
+
+  if (must_check_reply) {
+    unsigned long timer = millis() + 200;
+
+    while (!client_available(client)) {
+      if (timeOutReached(timer)) { return false; }
+      delay(1);
+    }
+
+    // Read all the lines of the reply from server and print them to Serial
+    while (client_available(client) && !success) {
+      //   String line = client.readStringUntil('\n');
+      String line;
+      safeReadStringUntil(client, line, '\n');
+
+#ifndef BUILD_NO_DEBUG
+
+      if (loglevelActiveFor(LOG_LEVEL_DEBUG_MORE)) {
+        if (line.length() > 80) {
+          addLog(LOG_LEVEL_DEBUG_MORE, line.substring(0, 80));
+        } else {
+          addLog(LOG_LEVEL_DEBUG_MORE, line);
+        }
+      }
+#endif // ifndef BUILD_NO_DEBUG
+
+      if (line.startsWith(F("HTTP/1.1 2")))
+      {
+        success = true;
+
+        // Leave this debug info in the build, regardless of the
+        // BUILD_NO_DEBUG flags.
+        if (loglevelActiveFor(LOG_LEVEL_DEBUG)) {
+          String log = F("HTTP : ");
+          log += logIdentifier;
+          log += F(" Success! ");
+          log += line;
+          addLog(LOG_LEVEL_DEBUG, log);
+        }
+      } else if (line.startsWith(F("HTTP/1.1 4"))) {
+        if (loglevelActiveFor(LOG_LEVEL_ERROR)) {
+          String log = F("HTTP : ");
+          log += logIdentifier;
+          log += F(" Error: ");
+          log += line;
+          addLog(LOG_LEVEL_ERROR, log);
+        }
+#ifndef BUILD_NO_DEBUG
+        addLog(LOG_LEVEL_DEBUG_MORE, postStr);
+#endif // ifndef BUILD_NO_DEBUG
+      }
+      delay(0);
+    }
+  }
+#ifndef BUILD_NO_DEBUG
+
+  if (loglevelActiveFor(LOG_LEVEL_DEBUG)) {
+    String log = F("HTTP : ");
+    log += logIdentifier;
+    log += F(" closing connection");
+    addLog(LOG_LEVEL_DEBUG, log);
+  }
+#endif // ifndef BUILD_NO_DEBUG
+
+  client.flush();
+  client.stop();
+  return success;
+}
+
+bool send_via_http(int controller_number, WiFiClient& client, const String& postStr, bool must_check_reply) {
+  return send_via_http(get_formatted_Controller_number(controller_number), client, postStr, must_check_reply);
+}
